Added gss codes for new councils, created new lookup from gss codes (old and new) to registry.
