+++ README.md
@@ -3,13 +3,14 @@ Lookup between the many, many different ways of naming and coding UK local autho
 
 If you ever try to match UK local authority statistics against each other you'll quickly find your efforts frustrated by inconsistent uses of official codes and different versions of LA names. These lookups are an ongoing tool to help translate between these. 
 
-* The uk_local_authorities.csv file contain the table to match between different codes. 
-* The uk_local_authority_messy_lookup.csv file combine all alternate names into one column to let you quickly match any data and get back a canonical code (which can be checked against the other table to convert to your preferred format). Also combined with this [name lookup](https://github.com/openregister/local-authority-data/edit/master/maps/name.tsv) from openregister for fuller coverage.
+* The uk_local_authorities.xlsx and .csv file contain the table to match between different codes - xlsx is the master, .csv is generated by build.py
+* The lookup_name_to_registry.csv file combine all alternate names into one column to let you quickly match any data and get back a canonical code (which can be checked against the other table to convert to your preferred format). Also combined with this [name lookup](https://github.com/openregister/local-authority-data/edit/master/maps/name.tsv) from openregister for fuller coverage.
+* The lookup_gss_to_registry.csv file converts from present and former GSS LAD codes (which refer to the boundary shape rather than the authority to the canonical code. 
 * lookups/lsoa_datazone_la.csv converts from LSOA/datazone for Scotland to the local-authority-code.
 
 This is using the official [local authority registers](https://github.com/openregister/local-authority-data) in development to provide a canonical three character code and a canonical name for the local authority. This is then expanded with current and older ONS codes and varying forms of the LA name (the real point of the exercise being to quickly get from horrible code-less data to other formats quickly).
 
-Useful fields:
+##Useful fields
 
 * local-authority-code - canonical three character code
 * official-name - canonical local authority name
@@ -26,7 +27,8 @@ Useful fields:
 * ofcom - variant on snac authority code used in coding survey responses (for instance in OFCOM's [Connected Nations report](https://www.ofcom.org.uk/research-and-data/infrastructure-research/connected-nations-2015))
 * ecode - used in [revenue and accounting documents](https://www.gov.uk/government/collections/local-authority-revenue-expenditure-and-financing)
 
-
 Additional maps can be found [here](https://github.com/openregister/local-authority-data/tree/master/maps).
 
 *The code used for the Greater London Authority is the gss code is for London region - when working from code point open use E18000007 and the NHS_HA_code column.
+
+Use build.py to recreate lookups and .csv version.
