+++ lookup_gss_to_registry.csv
@@ -1,0 +1,437 @@
+gss-code,local-authority-code
+E07000223,ADU
+E07000026,ALL
+E07000032,AMB
+E07000224,ARU
+E07000105,ASF
+E07000170,ASH
+E07000004,AYL
+E07000200,BAB
+E07000171,BAE
+E07000066,BAI
+E07000084,BAN
+E07000027,BAR
+E06000022,BAS
+E06000008,BBD
+E06000055,BDF
+E09000002,BDG
+E09000005,BEN
+E09000004,BEX
+E08000025,BIR
+E10000002,BKM
+E07000129,BLA
+E06000028,BMH
+E09000003,BNE
+E06000043,BNH
+E08000016,BNS
+E08000001,BOL
+E07000033,BOS
+E07000136,BOT
+E06000009,BPL
+E07000067,BRA
+E06000036,BRC
+E08000032,BRD
+E07000143,BRE
+E07000234,BRM
+E07000144,BRO
+E07000172,BRT
+E07000068,BRW
+E07000095,BRX
+E09000006,BRY
+E06000023,BST
+E07000117,BUN
+E08000002,BUR
+E07000008,CAB
+E10000003,CAM
+E07000192,CAN
+E07000028,CAR
+E07000069,CAS
+E07000106,CAT
+E06000056,CBF
+E07000130,CHA
+E07000048,CHC
+E06000049,CHE
+E07000225,CHI
+E07000070,CHL
+E07000005,CHN
+E07000118,CHO
+E07000177,CHR
+E07000034,CHS
+E07000078,CHT
+E06000050,CHW
+E08000033,CLD
+E10000006,CMA
+E09000007,CMD
+E07000071,COL
+E06000052,CON
+E07000029,COP
+E07000150,COR
+E07000079,COT
+E08000026,COV
+E07000163,CRA
+E07000226,CRW
+E09000008,CRY
+E07000096,DAC
+E06000005,DAL
+E07000107,DAR
+E07000151,DAV
+E10000007,DBY
+E07000035,DEB
+E06000015,DER
+E10000008,DEV
+E08000017,DNC
+E10000009,DOR
+E07000108,DOV
+E08000027,DUD
+E06000047,DUR
+E09000009,EAL
+E07000061,EAS
+E07000086,EAT
+E07000009,ECA
+E07000040,EDE
+E07000030,EDN
+E07000049,EDO
+E07000085,EHA
+E07000242,EHE
+E07000097,EHE
+E07000137,ELI
+E07000207,ELM
+E09000010,ENF
+E07000152,ENO
+E07000072,EPP
+E07000208,EPS
+E07000036,ERE
+E06000011,ERY
+E10000012,ESS
+E07000193,EST
+E10000011,ESX
+E07000041,EXE
+E07000087,FAR
+E07000010,FEN
+E07000080,FOE
+E07000201,FOR
+E07000119,FYL
+E08000037,GAT
+E08000020,GAT
+E07000173,GED
+E12000007,GLA
+E07000081,GLO
+E10000013,GLS
+E07000088,GOS
+E07000109,GRA
+E09000011,GRE
+E07000209,GRT
+E07000145,GRY
+E07000090,HAA
+E07000164,HAE
+E07000165,HAG
+E06000006,HAL
+E10000014,HAM
+E07000131,HAO
+E07000073,HAR
+E07000062,HAS
+E07000089,HAT
+E09000016,HAV
+E09000012,HCK
+E06000019,HEF
+E07000098,HER
+E07000037,HIG
+E09000017,HIL
+E07000132,HIN
+E09000013,HMF
+E09000018,HNS
+E07000227,HOR
+E06000001,HPL
+E10000015,HRT
+E09000015,HRW
+E09000014,HRY
+E07000011,HUN
+E07000120,HYN
+E06000053,IOS
+E06000046,IOW
+E07000202,IPS
+E09000019,ISL
+E09000020,KEC
+E10000016,KEN
+E07000153,KET
+E06000010,KHL
+E07000146,KIN
+E08000034,KIR
+E09000021,KTT
+E08000011,KWL
+E07000121,LAC
+E10000017,LAN
+E09000022,LBH
+E06000016,LCE
+E08000035,LDS
+E10000018,LEC
+E07000063,LEE
+E09000023,LEW
+E07000138,LIC
+E07000194,LIF
+E10000019,LIN
+E08000012,LIV
+E09000001,LND
+E06000032,LUT
+E07000110,MAI
+E07000074,MAL
+E08000003,MAN
+E07000174,MAS
+E07000235,MAV
+E06000002,MDB
+E07000042,MDE
+E06000035,MDW
+E07000133,MEL
+E07000187,MEN
+E06000042,MIK
+E07000210,MOL
+E09000024,MRT
+E07000228,MSS
+E07000203,MSU
+E06000057,NBL
+E06000048,NBL
+E07000043,NDE
+E07000050,NDO
+E07000175,NEA
+E07000195,NEC
+E07000038,NED
+E06000012,NEL
+E08000021,NET
+E07000091,NEW
+E10000020,NFK
+E06000018,NGM
+E07000099,NHE
+E07000139,NKE
+E06000013,NLN
+E07000147,NNO
+E07000154,NOR
+E07000148,NOW
+E06000024,NSM
+E10000021,NTH
+E10000024,NTT
+E08000022,NTY
+E07000219,NUN
+E07000218,NWA
+E07000134,NWL
+E09000025,NWM
+E10000023,NYK
+E07000135,OAD
+E08000004,OLD
+E10000025,OXF
+E07000178,OXO
+E07000122,PEN
+E06000026,PLY
+E06000029,POL
+E06000044,POR
+E07000123,PRE
+E06000031,PTE
+E07000051,PUR
+E06000003,RCC
+E08000005,RCH
+E09000026,RDB
+E06000038,RDG
+E07000236,RED
+E07000211,REI
+E07000124,RIB
+E09000027,RIC
+E07000166,RIH
+E07000075,ROC
+E07000064,ROH
+E07000125,ROS
+E08000018,ROT
+E07000220,RUG
+E07000092,RUH
+E07000212,RUN
+E07000176,RUS
+E06000017,RUT
+E07000167,RYE
+E07000240,SAL
+E07000100,SAL
+E08000028,SAW
+E07000006,SBU
+E07000012,SCA
+E07000168,SCE
+E07000039,SDE
+E07000204,SED
+E07000188,SEG
+E07000169,SEL
+E07000111,SEV
+E10000029,SFK
+E08000014,SFT
+E06000025,SGC
+E07000044,SHA
+E07000112,SHE
+E08000019,SHF
+E08000013,SHN
+E07000140,SHO
+E06000051,SHR
+E07000141,SKE
+E08000007,SKP
+E07000031,SLA
+E08000006,SLF
+E06000039,SLG
+E08000024,SND
+E07000149,SNO
+E07000155,SNR
+E08000029,SOL
+E10000027,SOM
+E06000033,SOS
+E07000179,SOX
+E07000213,SPE
+E07000126,SRI
+E10000030,SRY
+E07000189,SSO
+E07000196,SST
+E07000197,STA
+E06000021,STE
+E07000198,STF
+E06000045,STH
+E09000029,STN
+E07000082,STO
+E07000221,STR
+E10000028,STS
+E06000004,STT
+E07000243,STV
+E07000101,STV
+E08000023,STY
+E07000205,SUF
+E07000214,SUR
+E06000030,SWD
+E09000028,SWK
+E07000113,SWL
+E08000008,TAM
+E07000215,TAN
+E07000190,TAU
+E07000199,TAW
+E07000045,TEI
+E07000076,TEN
+E07000093,TES
+E07000083,TEW
+E06000020,TFW
+E07000114,THA
+E07000102,THE
+E06000034,THR
+E06000027,TOB
+E07000115,TON
+E07000046,TOR
+E08000009,TRF
+E07000116,TUN
+E09000030,TWH
+E07000077,UTT
+E07000180,VAL
+E07000216,WAE
+E10000031,WAR
+E07000103,WAT
+E07000206,WAV
+E07000222,WAW
+E06000037,WBK
+E07000047,WDE
+E07000052,WDO
+E07000065,WEA
+E07000156,WEL
+E07000241,WEW
+E07000104,WEW
+E07000053,WEY
+E09000031,WFT
+E08000010,WGN
+E06000054,WIL
+E07000094,WIN
+E08000036,WKF
+E07000127,WLA
+E07000142,WLI
+E08000030,WLL
+E08000031,WLV
+E09000032,WND
+E06000040,WNM
+E07000237,WOC
+E07000217,WOI
+E06000041,WOK
+E10000034,WOR
+E07000229,WOT
+E07000181,WOX
+E08000015,WRL
+E06000007,WRT
+E09000033,WSM
+E07000191,WSO
+E10000032,WSX
+E07000238,WYC
+E07000239,WYE
+E07000007,WYO
+E07000128,WYR
+E06000014,YOR
+N09000002,ABC
+N09000011,AND
+N09000001,ANN
+N09000003,BFS
+N09000004,CCG
+N09000005,DRS
+N09000006,FMO
+N09000007,LBC
+N09000008,MEA
+N09000009,MUL
+N09000010,NMD
+S12000034,ABD
+S12000033,ABE
+S12000035,AGB
+S12000041,ANS
+S12000005,CLK
+S12000006,DGY
+S12000042,DND
+S12000008,EAY
+S12000036,EDH
+S12000045,EDU
+S12000009,EDU
+S12000010,ELN
+S12000013,ELS
+S12000011,ERW
+S12000014,FAL
+S12000047,FIF
+S12000015,FIF
+S12000049,GLG
+S12000046,GLG
+S12000043,GLG
+S12000017,HLD
+S12000018,IVC
+S12000019,MLN
+S12000020,MRY
+S12000021,NAY
+S12000050,NLK
+S12000044,NLK
+S12000023,ORK
+S12000048,PKN
+S12000024,PKN
+S12000038,RFW
+S12000028,SAY
+S12000026,SCB
+S12000029,SLK
+S12000030,STG
+S12000039,WDU
+S12000040,WLN
+S12000027,ZET
+W06000001,AGY
+W06000013,BGE
+W06000019,BGW
+W06000018,CAY
+W06000008,CGN
+W06000010,CMN
+W06000015,CRF
+W06000003,CWY
+W06000004,DEN
+W06000005,FLN
+W06000002,GWN
+W06000021,MON
+W06000024,MTY
+W06000012,NTL
+W06000022,NWP
+W06000009,PEM
+W06000023,POW
+W06000016,RCT
+W06000011,SWA
+W06000020,TOF
+W06000014,VGL
+W06000006,WRX
+E06000058,BPC
+E06000059,DST
+E07000246,SWT
+E07000245,WSK
+E07000244,ESK
